authorization

+++ controllers/note.js
@@ -66,7 +66,7 @@ exports.getNote = (req, res, next) => {
   const { id } = req.params;
 
   Note.findById(id)
-    .populate("author", "username")
+    .populate("author", "username email")
     .then((note) => {
       res.status(200).json(note);
     })
@@ -82,6 +82,9 @@ exports.deleteNote = (req, res, next) => {
   const { id } = req.params;
   Note.findById(id)
     .then((note) => {
+      if (note.author.toString() !== req.userId) {
+        return res.status(401).json("Auth failed!");
+      }
       if (note.cover_image) {
         unlink(note.cover_image);
       }
@@ -104,6 +107,9 @@ exports.getOldNote = (req, res, next) => {
   const { id } = req.params;
   Note.findById(id)
     .then((note) => {
+       if (note.author.toString() !== req.userId) {
+         return res.status(401).json("Auth failed!");
+       }
       return res.status(200).json(note);
     })
     .catch((err) => {
@@ -120,6 +126,9 @@ exports.updateNote = (req, res, next) => {
 
   Note.findById(note_id)
     .then((note) => {
+       if (note.author.toString() !== req.userId) {
+         return res.status(401).json("Auth failed!");
+       }
       note.title = title;
       note.content = content;
       if (cover_image) {
